removed bad server.js file
