+++ controllers/api/userRoutes.js
@@ -43,27 +43,27 @@ router.post('/logout', (req, res) => {
   }
 });
 
-router.post('/create-account', async (req, res) => {
-  try {
+// router.post('/create-account', async (req, res) => {
+//   try {
     
-    const newUser = {
-      username: req.body.username,
-    };
+//     const newUser = {
+//       username: req.body.username,
+//     };
 
-    newUser.password = await bcrypt.hash(req.body.password, 10);
-    // create the newUser with the hashed password and save to DB
-    const userData = await User.create(newUser);
+//     newUser.password = await bcrypt.hash(req.body.password, 10);
+//     // create the newUser with the hashed password and save to DB
+//     const userData = await User.create(newUser);
 
-    req.session.save(() => {
-      req.session.user_id = userData.id;
-      req.session.logged_in = true;
+//     req.session.save(() => {
+//       req.session.user_id = userData.id;
+//       req.session.logged_in = true;
       
-      res.json({ user: userData, message: 'Created new account. You are now logged in!' });
-    });
+//       res.json({ user: userData, message: 'Created new account. You are now logged in!' });
+//     });
 
-  } catch (err) {
-    res.status(400).json(err);
-  }
-});
+//   } catch (err) {
+//     res.status(400).json(err);
+//   }
+// });
 
 module.exports = router;
