+++ controllers/homeRoutes.js
@@ -1,37 +1,38 @@
+const router = require('express').Router();
+const { User } = require('../models');
+const withAuth = require('../utils/auth');
 
+router.get('/', withAuth, async (req, res) => {
+  try {
+    const userData = await User.findAll({
+      attributes: { exclude: ['password'] },
+      order: [['name', 'ASC']],
+    });
+    console.log(userData)
 
+    const users = userData.map((project) => project.get({ plain: true }));
 
+    res.render('homepage', {
+      users,
+      logged_in: req.session.logged_in,
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
 
+router.get('/login', (req, res) => {
+  if (req.session.logged_in) {
+    res.redirect('/');
+    return;
+  }
 
-// const router = require('express').Router();
-// const { User } = require('../models');
-// const withAuth = require('../utils/auth');
+  res.render('login');
+});
 
-// router.get('/', withAuth, async (req, res) => {
-//   try {
-//     const userData = await User.findAll({
-//       attributes: { exclude: ['password'] },
-//       order: [['name', 'ASC']],
-//     });
+// router.get('/createAccount', (req, res) => {
+//   res.json()
+//   // res.render('createUser')
+// })
 
-//     const users = userData.map((project) => project.get({ plain: true }));
-
-//     res.render('homepage', {
-//       users,
-//       logged_in: req.session.logged_in,
-//     });
-//   } catch (err) {
-//     res.status(500).json(err);
-//   }
-// });
-
-// router.get('/login', (req, res) => {
-//   if (req.session.logged_in) {
-//     res.redirect('/');
-//     return;
-//   }
-
-//   res.render('login');
-// });
-
-// module.exports = router;
+module.exports = router;
