+++ seeds/userData.json
@@ -16,7 +16,7 @@
     "password": "password12345"
   },
   {
-    "name": "Blake",
+    "username": "Blake",
     "password": "password12345"
   }
 ]
