+++ views/login.handlebars
@@ -14,6 +14,6 @@
   </div>
 </form>
 
-<button class="btn btn-primary">Create new account</button>
+<button class="btn btn-primary create-account" type="submit">Create new account</button>
 
 <script src="/js/login.js"></script>
